rename fused_quant into fused_mxfp4_quant
